requirements.txt
torch>=2.1
torchvision>=0.16
torchcam>=0.4.0
opencv-python
pillow
streamlit
tqdm
